Revert "Updated Spark Examples to latest images" (#115)
