--- Data-Analytics/Spark/DataProcessTransfer/DataProcessTransferFts.yaml
@@ -1,44 +1,0 @@
-apiVersion: "sparkoperator.hpe.com/v1beta2"
-kind: SparkApplication
-metadata:
-  name: data-process-transfer-demo
-  namespace: <username>
-spec:
-  sparkConf:
-    spark.hadoop.fs.s3a.path.style.access: "true"
-    spark.hadoop.fs.s3a.impl: "org.apache.hadoop.fs.s3a.S3AFileSystem"
-    spark.hadoop.fs.s3a.connection.ssl.enabled: "false"
-    spark.hadoop.fs.s3a.aws.credentials.provider: "org.apache.spark.s3a.EzSparkAWSCredentialProvider"
-    spark.hadoop.fs.s3a.endpoint: http://local-s3-service.ezdata-system.svc.cluster.local:30000
-  type: Scala
-  sparkVersion: 3.5.0
-  mode: cluster
-  image: gcr.io/mapr-252711/spark-3.5.0:v3.5.0
-  imagePullPolicy: Always
-  mainApplicationFile: "local:///mounts/shared-volume/ezua-tutorials/Data-Analytics/Spark/DataProcessTransfer/DataProcessTransfer.jar"
-  mainClass: com.mapr.sparkdemo.DataProcessTransfer
-  arguments:
-    - s3a://ezua-demo/data/financial.csv               # source path
-    - csv                                              # source format
-    - file:///mounts/shared-volume/financial-processed # destination path
-    - parquet                                          # destination format
-  restartPolicy:
-    type: Never
-  imagePullSecrets:
-    - imagepull
-  driver:
-    labels:
-      version: 3.5.0
-    volumeMounts:
-      - name: shared-volume
-        mountPath: /mounts/shared-volume
-  executor:
-    labels:
-      version: 3.5.0
-    volumeMounts:
-      - name: shared-volume
-        mountPath: /mounts/shared-volume
-  volumes:
-    - name: shared-volume
-      persistentVolumeClaim:
-        claimName: kubeflow-shared-pvc
