+++ Data-Analytics/Spark/README.md
@@ -1,9 +1,15 @@
 # EZAF Spark Example
 
-This folder contains Spark examples which you can manually submit via Spark operator in your **k8s cluster** .
+This is a Spark ETL example which you can manually submit via Spark operator in your **k8s cluster** .
+Deployed Spark app reads the raw data from **S3 Object Store**, preprocesses it in distributed mode and
+writes to the **[HPE Data Fabric](https://www.hpe.com/us/en/software/ezmeral-data-fabric.html).**
+Processed data will be stored to **PV**, referenced by **PVC**. With that,
+you can access it from other EZAF components by plugging the same **PVC**
 
 ### Configuration
 
+To configure Spark example, we use [Embed .jar file into the **docker image**](k8s/DataProcessTransferFts-JarLocal-3.4.0.yaml)
+of delivering app's **.jar** file.
 
 > **Note:** For more detailed acquaintance with Spark config you can look through following table with more extended explanation
 
@@ -12,8 +18,8 @@ This folder contains Spark examples which you can manually submit via Spark oper
 
 | Parameter               | Description                                                                                            | Example Value                                                                                                                                                                                                                                                                                                                                          |
 |-------------------------|--------------------------------------------------------------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| `*.image`               | Image that is used to run Spark itself                                                                 | <ul><li>`gcr.io/mapr-252711/spark-3.5.0:v3.5.0`</li>Pure Spark image<ul>                                                                                                                                                              |
- | `*.mainApplicationFile` | Path inside the pod that is used by spark to find `.jar` file with app                                 | <ul><li>`local:///tmp/<Application>.jar`</li>Path in which `.jar` file was embedded into image<br/><br/><li>`local///mounts/<Application>.jar`</li>:warning: The specified path must be mounted to the `PV`<br/><br/><li>`maprfs:///home/spark/<Application>.jar`</li>Path in which `.jar` file can be reached im `MaprFS`</ul> |
+| `*.image`               | Image that is used to run Spark itself                                                                 | <ul><li>`gcr.io/mapr-252711/ezaf-spark-demo-example:fy23-q3-fts`</li>Custom image with `.jar` file embedded<br/><br/><li>`gcr.io/mapr-252711/spark-3.4.0:v3.4.0`</li>Pure Spark image<ul>                                                                                                                                                              |
+ | `*.mainApplicationFile` | Path inside the pod that is used by spark to find `.jar` file with app                                 | <ul><li>`local:///tmp/DataProcessTransfer.jar`</li>Path in which `.jar` file was embedded into image<br/><br/><li>`local///mounts/data/DataProcessTransfer.jar`</li>:warning: The specified path must be mounted to the `PV`<br/><br/><li>`maprfs:///home/spark/DataProcessTransfer.jar`</li>Path in which `.jar` file can be reached im `MaprFS`</ul> |
  | `*.mainClass`           | The fully qualified name of the class that contains the main method for the Java and Scala application | `com.mapr.sparkdemo.DataTransferDemo`                                                                                                                                                                                                                                                                                                                  |
  | `*.arguments`           | Arguments passed to the main method of your main class (_i.e._ command arguments)                      | <ul><li>`s3a://ezaf-demo/data/financial.csv`</li>Data source path<br/><br/><li>`csv`</li>Data source format<br/><br/><li>`file:///mounts/data/financial-processed`<br/>Data destination path. :warning: The specified path must be mounted to `PV`</li><br/><li>`parquet`</li>Data destination format</ul>                                             | 
  | `*.mountPath`           | Path inside the Pod that would be mounted to persistent storage                                        | `/mounts/data`                                                                                                                                                                                                                                                                                                                                         |
@@ -22,22 +28,37 @@ This folder contains Spark examples which you can manually submit via Spark oper
 
 </details>
 
+### Installation
+
+Spark App can be deployed via [installation script][1].
 
 #### Usage
-By default, the workflow consists of these steps:
 
-1. You create a SparkApplication object in k8s by applying the object manifest in yaml file
+```bash
+/bin/bash devops/install.sh
+```
 
-2. Autotix webhook intercepts ‘create sparkapp’ request, mutates and validates it
+By default, its workflow consists of 4 steps:
 
-3. Mutated and validated object is stored within k8s etcd
+1. Gets [.jar][2] file, based on [scala code](src/DataProcessTransfer/DataProcessTransfer.scala)
 
-4. Spark operator finds new object in etcd and starts processing: it parses the object and creates a ‘spark submitter’ job.
+2. Inserts the [.jar][2] file to the custom [image](dockerfiles/SparkJarLocal-3.4.0.Dockerfile).
 
-5. Submitter job performs actual ‘spark-submit’ call, which in turn creates driver pod
+3. Pushes built image to the repo, specified in [script][1]
 
-6. Spark operator internal pod webhook intercepts the driver pod creation and applies pod mutations if needed (e.g., it mounts secrets and PVCs). Mutated pod spec is stored within etcd
+4. Deploys your Spark app as the LDAP user, specified in [script][1]
 
-7. Pod controller creates driver pod
 
-8. Same happens with executors
+### Troubleshooting
+
+1. You cannot submit your app due to such an error:
+`Please verify the user is in LDAP and has rights to use the Tenant` on the applying yaml step
+   * Tip: run `kubectl -n [namespace] describe rolebindings.rbac.authorization.k8s.io hpe-[namespace]-user` 
+    and verify that user (or user group) is in the subject list.
+2. Your application fails with such an error: 
+`ERROR FileOutputCommitter: Mkdirs failed to create [SomePath]`
+    * Tip: Make sure destination path is available for writing within the context of Linux file permissions in `maprfs`
+
+
+[1]:devops/install.sh
+[2]:src/DataProcessTransfer/DataProcessTransfer.jar
