+++ Data-Analytics/Spark/dockerfiles/SparkJarLocal-3.3.1.Dockerfile
@@ -1,0 +1,6 @@
+FROM gcr.io/mapr-252711/spark-3.3.1:v3.3.1
+
+ARG SPARK_APP_SOURCE
+ARG SPARK_APP_DEST
+
+COPY $SPARK_APP_SOURCE $SPARK_APP_DEST
