+++ Data-Analytics/Spark/dockerfiles/SparkJarLocal-3.4.0.Dockerfile
@@ -1,0 +1,6 @@
+FROM gcr.io/mapr-252711/spark-3.4.0:v3.4.0
+
+ARG SPARK_APP_SOURCE
+ARG SPARK_APP_DEST
+
+COPY $SPARK_APP_SOURCE $SPARK_APP_DEST
