--- Data-Analytics/Spark/pi/README.md
@@ -1,14 +1,0 @@
-# EZAF Spark Example
-
-This folder contains Spark Pi example which you can manually submit via Spark operator in your **k8s cluster** .
-
-#### Usage
-By default, the workflow consists of these steps:
-
-1. Navigate to the Spark Operator under Analytics in Ezmeral Unified Analytics.
-
-2. Create Application by uploading the `pi.yaml` file.
-
-3. Review the entire yaml for any discrepancy and click on Create Spark Application.
-
-4. Watch the logs for expected output, it would print the approximate value of pi.
