--- Data-Analytics/Spark/pi/pi.yaml
@@ -1,32 +1,0 @@
-apiVersion: "sparkoperator.hpe.com/v1beta2"
-kind: SparkApplication
-metadata:
-  name: spark-pi
-  namespace: <username>
-spec:
-  type: Python
-  sparkVersion: 3.5.0
-  mode: cluster
-  image: gcr.io/mapr-252711/spark-3.5.0:v3.5.0
-  imagePullPolicy: Always
-  mainApplicationFile: "local:///mounts/shared-volume/ezua-tutorials/Data-Analytics/Spark/pi/pi.py"
-  restartPolicy:
-    type: Never
-  imagePullSecrets:
-    - imagepull
-  driver:
-    labels:
-      version: 3.5.0
-    volumeMounts:
-      - name: shared-volume
-        mountPath: /mounts/shared-volume
-  executor:
-    labels:
-      version: 3.5.0
-    volumeMounts:
-      - name: shared-volume
-        mountPath: /mounts/shared-volume
-  volumes:
-    - name: shared-volume
-      persistentVolumeClaim:
-        claimName: kubeflow-shared-pvc
