--- Data-Analytics/Spark/wordcount/README.md
@@ -1,14 +1,0 @@
-# EZAF Spark Example
-
-This folder contains Word Count Spark example which you can manually submit via Spark operator in your **k8s cluster** .
-
-#### Usage
-By default, the workflow consists of these steps:
-
-1. Navigate to the Spark Operator under Analytics in Ezmeral Unified Analytics.
-
-2. Create Application by uploading the `wordcount.yaml` file.
-
-3. Review the entire yaml for any discrepancy and click on Create Spark Application.
-
-4. Watch the logs fpr expected output, it print the output of each word and it's count.
